refactor: combine tracing & metrics in observability package
