+++ internal/api/metrics/no_metrics.go
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-//go:build nometrics
+//go:build nootel
 
 package metrics
